Move content under public routes
