+++ src/app/(public)/[locale]/content/_components/ContentCoverImage.tsx
