+++ src/app/(public)/[locale]/content/_components/content-cover-image.module.scss
