+++ src/app/(public)/[locale]/content/contact-us/page.tsx
@@ -1,4 +1,4 @@
-import ContentPage from "../_components/ContentPage";
+import ContentPage from "../../../../(public)/[locale]/content/_components/ContentPage";
 import { getTranslations, unstable_setRequestLocale } from "next-intl/server";
 import { ALL_LOCALES } from "@/i18n/config";
 import { metaTitle } from "@/lib/metadataUtils";
@@ -9,11 +9,11 @@ export function generateStaticParams() {
 }
 
 export async function generateMetadata({
-  params,
+  params: { locale },
 }: {
   params: { locale: string };
 }) {
-  const t = await getTranslations("home");
+  const t = await getTranslations({ locale, namespace: "home" });
   return {
     title: metaTitle(t("contact.title")),
   };
